% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbrefscraper-package.R
\docType{package}
\name{bbrefscraper-package}
\alias{bbrefscraper}
\alias{bbrefscraper-package}
\title{bbrefscraper: Scraper for basketball-reference.com}
\description{
This package enables an easy interface from R
    to scrape data from different parts of basketball-reference.com
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tomiaJO/bbrefscraper}
}

}
\author{
\strong{Maintainer}: Tamas Koncz \email{t.koncz@gmail.com}

}
\keyword{internal}
